% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_aggregate.R
\name{aggregate_fp}
\alias{aggregate_fp}
\title{aggregate_fp}
\usage{
aggregate_fp(posterior_samples, population_data)
}
\arguments{
\item{posterior_samples}{}

\item{population_data}{}
}
\value{

}
\description{
aggregate_fp
}
