% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_aggregate.R
\name{aggregate_fp_multilevel}
\alias{aggregate_fp_multilevel}
\title{aggregate_fp_multilevel}
\usage{
aggregate_fp_multilevel(
  posterior_samples,
  division_level_data,
  population_data
)
}
\arguments{
\item{posterior_samples}{posteriors samples from \code{\link{posterior_samples_array_format}}}

\item{division_level_data}{data frame with country codes and corresponding aggregate level}

\item{population_data}{population data, union and years of the samples}
}
\description{
aggregate_fp_multilevel
}
\examples{
dimnames(posterior_samples) <- list(division_numeric_code, NULL, NULL, NULL) #provide corresponding division numeric codes in posterior_samples attributes
population_data <- population_counts \%>\%
  dplyr::filter(is_in_union == union) \%>\%
  dplyr::filter(mid_year <= last_year) \%>\%
  dplyr::filter(mid_year >= first_year)
division_level_data <- divisions \%>\%
  mutate(division_level = region_numeric_code)\%>\%
  select(division_numeric_code, division_level)
posterior_samples_list <- weight_samples(division_level_data, population_data, posterior_samples)

}
