% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_aggregate.R
\name{aggregate_fp_sub}
\alias{aggregate_fp_sub}
\title{aggregate_fp_sub}
\usage{
aggregate_fp_sub(posterior_samples, weight_df)
}
\arguments{
\item{weights}{\emph{'Vector'} A vector of weights selected from \code{\link[weight_generator]{weight_generator}}}
}
\value{
\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions
}
\description{
aggregate_fp_sub
}
