% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_summarise.R
\name{array_summarise}
\alias{array_summarise}
\title{array_summarise}
\usage{
array_summarise(x, h = standard_summary, ...)
}
\arguments{
\item{x}{an array of samples with margins:
1) chain
2) iteration
4) (year) optionally
5) indicator}

\item{h}{a function that takes a  vector of samples and
returns one or more summaries of the vector.}
}
\value{
a set of summaries with margins:
          1) (year) optionally
          2) summaries
}
\description{
Summarize indicator array for a single unit
}
