% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_aggregate.R
\name{calc_fp_aggregate}
\alias{calc_fp_aggregate}
\title{calc_fp_aggregate}
\usage{
calc_fp_aggregate(fits, population_data = NULL)
}
\arguments{
\item{fits}{\emph{\sQuote{List}} The list returned from \code{\link{fit_fp_csub}}}

\item{population_data}{\emph{\sQuote{Data.frame}} Population count data such as \code{\link{population_counts}}.}
}
\value{

}
\description{
calc_fp_aggregate
}
