% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_c.R
\name{calc_fp_c_autosave}
\alias{calc_fp_c_autosave}
\title{Calculate fp indicators from samples, autosave}
\usage{
calc_fp_c_autosave(runname = NULL, population_data = NULL)
}
\arguments{
\item{population_data}{\emph{\sQuote{Data.frame}} Population count data such as \code{\link{population_counts}}.}
}
\value{
\emph{NULL}
}
\description{
A wrapper for \code{\link{calc_fp_c}} which reads in the model fit from the respective `runname`. The output is saved in the results directory with the same `runname`. The output can be automatically read in by proceeding wrapper functions with the `runname` argument specified.
}
