% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_data.R
\name{core_data}
\alias{core_data}
\title{core_data}
\usage{
core_data(
  is_in_union,
  surveydata_filepath,
  division_numeric_code,
  first_year,
  last_year,
  subnational
)
}
\arguments{
\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}

\item{surveydata_filepath}{\emph{\sQuote{Character}} Path to survey data. Survey data should be a .csv with the following format \code{\link{contraceptive_use}}.}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}

\item{last_year}{\emph{\sQuote{Numeric}} The last year of model estimates in output. The model will be fit to all data, including dates after this date if available.}

\item{subnational}{'\emph{\sQuote{Logical}} If TRUE runs the sub national model.}
}
\description{
Creates core_data that consists of manipulated observations and run settings for \code{\link{fit_fp_c}}
}
