% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_series_type_relabel.R
\name{data_series_type_relabel}
\alias{data_series_type_relabel}
\title{index_datatype_supplement}
\usage{
data_series_type_relabel(data)
}
\arguments{
\item{data}{\emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link[core_data]{core_data}}}
}
\value{
data  \emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link[core_data]{core_data}}
}
\description{
index_datatype_supplement
}
