% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{divisions}
\alias{divisions}
\title{Divisions, unit data}
\format{
A data frame with 13 variables:
\describe{
  \item{division_numeric_code}{\emph{\sQuote{Numeric}} 3-digit \emph{M49} country code (\cite{United Nations, 2019}).}
  \item{name_country}{\emph{\sQuote{Character}}, country name.}
  \item{name_region}{\emph{\sQuote{Character}}, region name (e.g., \dQuote{Africa}).}
  \item{name_sub_region}{\emph{\sQuote{Character}}, subregion name (e.g., \dQuote{Eastern Africa}).}
  \item{region_numeric_code}{\emph{\sQuote{Numeric}} 3-digit \emph{M49} region code (\cite{United Nations, 2019}).}
  \item{sub_region_numeric_code}{\emph{\sQuote{Numeric}} 3-digit \emph{M49} subregion code (\cite{United Nations, 2019}).}
  \item{is_developed_region}{\emph{\sQuote{Character}} taking values \dQuote{Y} and \dQuote{N}. Indicates countries classified as belonging to developed regions (\code{is_developed_region == "Y"}).}
  \item{is_less_developed_region}{\emph{\sQuote{Character}} taking values \dQuote{Y} and \dQuote{N}. Indicates countries classified as belonging to less developed regions (\code{is_less_developed_region == "Y"}).}
  \item{is_least_developed_country}{\emph{\sQuote{Character}} taking values \dQuote{Y} and \dQuote{N}. Indicates countries classified as least developed (\code{is_least_developed_country == "Y"}).}
  \item{is_in_sub_saharan_africa}{\emph{\sQuote{Character}} taking values \dQuote{Y} and \dQuote{N}. Indicates countries in sub-Saharan Africa (\code{is_in_sub_saharan_africa == "Y"}).}
  \item{is_unmarried_sexual_activity}{\emph{\sQuote{Character}} taking values \dQuote{Y} and \dQuote{N}. Indicates countries classified as having low sexual activity among unmarried women (\code{is_unmarried_sexual_activity == "Y"}; see \cite{Wheldon et al., 2018}).}
  \item{is_low_population}{\emph{\sQuote{Character}} taking values \dQuote{Y} and \dQuote{N}. Indicates countries with population below 90,000 (\code{is_low_population == "Y"}).}
  \item{is_fp2020}{\emph{\sQuote{Character}} taking values \dQuote{Y} and \dQuote{N}. Indicates countries belonging to the group of \emph{FP2020} focus countries (\code{is_fp2020 == "Y"}; \cite{FP2020, 2019}).}
}
}
\source{
UN 2018,	2018.0.1
}
\usage{
divisions
}
\description{
A dataset containing units compiled by the UNPD
}
\references{
FP2020 (2019). \url{http://www.familyplanning2020.org/} (visited on 2019-10-21).

United Nations (2019). \emph{Standard Country or Area Codes for
Statistical Use (M49)}. United Nations, Department of Economic and
Social Affairs, Statistics
Division. \url{https://unstats.un.org/unsd/methodology/m49/}
(visited on 2019-10-21).

Wheldon, M. C., Kantorova, V., Ueffing, P., Dasgupta,
A. N. Z. (2018). Methods for estimating and projecting key family
planning indicators among all women of reproductive age. Technical
Paper No. 2018/2, United Nations, Department of Economic and Social Affairs, Population Division.
\url{https://www.un.org/en/development/desa/population/publications/pdf/technical/TP2018-2.pdf} (visited on 2019-10-21).
}
\keyword{datasets}
