% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{expand_contraceptive_use}
\alias{expand_contraceptive_use}
\title{Expand contraceptive use data to include derived values.}
\usage{
expand_contraceptive_use(contraceptive_use, simplify_logicals = TRUE)
}
\arguments{
\item{contraceptive_use}{`data.frame` Contraceptive use data.}

\item{simplify_logicals}{`logical` Convert Y/N to TRUE/FALSE.}
}
\value{
`data.frame` Expanded contraceptive use data..
}
\description{
Expand contraceptive use data to include derived values.
}
