% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fp_c.R
\name{fit_fp_c}
\alias{fit_fp_c}
\title{Fit one country model}
\usage{
fit_fp_c(population_data = NULL, is_in_union, ...)
}
\arguments{
\item{population_data}{\emph{\sQuote{Data.frame}} Population count data such as \code{\link{population_counts}}.}

\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}

\item{last_year}{\emph{\sQuote{Numeric}} The last year of model estimates in output. The model will be fit to all data, including dates after this date if available.}

\item{surveydata_filepath}{\emph{\sQuote{Character}} Path to survey data. Survey data should be a .csv with the following format \code{\link{contraceptive_use}}.}

\item{service_stats_filepath}{\emph{\sQuote{Character}}}

\item{subnational}{'\emph{\sQuote{Logical}} If TRUE runs the sub national model.}

\item{diagnostic}{'\emph{\sQuote{Logical}} If TRUE saves the full jags output.}
}
\value{
\emph{\sQuote{List}}
\enumerate{
  \item \strong{posterior_samples}  \emph{\sQuote{Numeric array}} An array of samples of dimension chains x samples x years x proportions.
  \item \strong{core_data}          \emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link{core_data}}.
}
}
\description{
Fits the family planning estimation model and returns samples and bias adjusted observations for respective run. Please see \code{\link{fit_fp_csub}} for additional argument default settings
}
\examples{
See the reposiotry url in references for detailed examples
}
\references{
\url{https://github.com/FPcounts/FPEM}
}
