% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fp_c.R
\name{fit_fp_c_autosave}
\alias{fit_fp_c_autosave}
\title{Fit one country model and save output}
\usage{
fit_fp_c_autosave(runname = NULL, ...)
}
\arguments{
\item{runname}{\emph{\sQuote{Character}} The name given to the file generated by this wrapper.}

\item{...}{Please see \code{\link{fit_fp_csub}} for additional arguments.}
}
\description{
A wrapper for \code{\link{fit_fp_csub}} augmented to save the output
}
