% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_check.R
\name{format_check}
\alias{format_check}
\title{format check}
\usage{
format_check(format_list, data)
}
\arguments{
\item{format_list}{}

\item{data}{}
}
\value{
an informative error message if data does not pass validation
}
\description{
format check
}
