% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_contraceptive_use_modern}
\alias{get_contraceptive_use_modern}
\title{Get modern CPR}
\usage{
get_contraceptive_use_modern(posterior_samples, first_year)
}
\arguments{
\item{posterior_samples}{\emph{\sQuote{Array}} The samples array from \code{\link{fit_fp_csub}}.}

\item{first_year}{`integer` Earliest year represented in the data}
}
\value{
`data.frame` Values by year and percentile
}
\description{
Get modern CPR
}
