% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_demand_satisfied_modern}
\alias{get_demand_satisfied_modern}
\title{Get demand satisfied with a modern method}
\usage{
get_demand_satisfied_modern(posterior_samples, first_year)
}
\arguments{
\item{posterior_samples}{\emph{\sQuote{Array}} The samples array from \code{\link{fit_fp_csub}}.}

\item{first_year}{`integer` Earliest year represented in the data}
}
\value{
`data.frame` Values by year and percentile
}
\description{
`demand_satisfied_modern = modern_cpr / demand`
}
