% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_posterior_probability_from_cutoff_target}
\alias{get_posterior_probability_from_cutoff_target}
\title{Get posterior probability from proportion or population}
\usage{
get_posterior_probability_from_cutoff_target(
  posterior_samples,
  population_count_year = 1,
  population_count_relative_to_year = 1,
  indicator,
  year,
  relative_to_year = NULL,
  above,
  cutoff
)
}
\arguments{
\item{posterior_samples}{\emph{\sQuote{Array}} The samples array from \code{\link{fit_fp_csub}}.}

\item{indicator}{`integer` Indicator index (1 = modern, 2 = traditional, 3 = unmet)}

\item{year}{`integer` Year index}

\item{above}{`logical` If FALSE then it's below}

\item{cutoff}{`numeric` Cutoff proportion or population}

\item{population_count}{`integer` Number of individuals in the sample population (1 to calculate from proportion)}
}
\value{
`numeric` Posterior probability
}
\description{
Get posterior probability from proportion or population
}
