% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_progress}
\alias{get_progress}
\title{Get progress for indicator}
\usage{
get_progress(
  posterior_samples,
  from_year_population_count = 1,
  to_year_population_count = 1,
  indicator,
  from_year,
  to_year,
  absolute = FALSE
)
}
\arguments{
\item{posterior_samples}{\emph{\sQuote{Array}} The samples array from \code{\link{fit_fp_csub}}.}

\item{from_year_population_count}{`integer`
Number of individuals in the sample population for "from" year (1 to calculate proportion)}

\item{to_year_population_count}{`integer`
Number of individuals in the sample population for "to" year (1 to calculate proportion)}

\item{indicator}{`integer` Indicator index (1 = modern, 2 = traditional, 3 = unmet)}

\item{from_year}{`integer` First year}

\item{to_year}{`integer` Last year}
}
\value{
`numeric` Median and 2.5%/97.5% quantiles of change
}
\description{
Get progress for indicator
}
