% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_target_from_posterior_probability}
\alias{get_target_from_posterior_probability}
\title{Get proportion or population from posterior probability}
\usage{
get_target_from_posterior_probability(
  posterior_samples,
  population_count_year = 1,
  population_count_relative_to_year = 1,
  indicator,
  year,
  relative_to_year = NULL,
  greater_than,
  probability
)
}
\arguments{
\item{posterior_samples}{\emph{\sQuote{Array}} The samples array from \code{\link{fit_fp_csub}}.}

\item{indicator}{`integer` Indicator index (1 = modern, 2 = traditional, 3 = unmet)}

\item{year}{`integer` Year index}

\item{greater_than}{`logical` If FALSE then it's less than}

\item{probability}{`numeric` Posterior probability}

\item{population_count}{`integer` Number of individuals in the sample population (1 to calculate proportion)}
}
\value{
`numeric` Proportion or population
}
\description{
Get proportion or population from posterior probability
}
