% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_se.R
\name{impute_indicator}
\alias{impute_indicator}
\title{Impute indicator}
\usage{
impute_indicator(data)
}
\arguments{
\item{wide_data}{\emph{'Data.frame'} A data.frame from \code{\link[contrapcetive_use]{contrapcetive_use}}}
}
\value{
\emph{'Data.frame'} A data.frame with logical indicator TRUE indicating it is imputed FALSE being a core value
}
\description{
Adds indicator columns to indicate if a value has been imputed
}
