% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_se.R
\name{impute_packagedata_se}
\alias{impute_packagedata_se}
\title{impute_packagedata_se}
\usage{
impute_packagedata_se(contraceptive_use)
}
\arguments{
\item{contraceptive_use}{}
}
\value{

}
\description{
Impute package data. This can be used on user data but takes a while to run. Instead impute_user_se uses pre-computed imputations for faster results.
}
