% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_se.R
\name{impute_packagedata_se_one_union}
\alias{impute_packagedata_se_one_union}
\title{impute_packagedata_se}
\usage{
impute_packagedata_se_one_union(contraceptive_use, is_in_union)
}
\arguments{
\item{contraceptive_use}{}

\item{is_in_union}{}
}
\value{

}
\description{
Impute package data per union. This can be used on user data but takes a while to run. Instead impute_user_se uses pre-computed imputations for faster results.
}
