% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_se.R
\name{impute_user_se}
\alias{impute_user_se}
\title{Impute user survey data standard errors}
\usage{
impute_user_se(user_data, subnational, is_in_union)
}
\arguments{
\item{user_data}{\emph{\sQuote{Data.frame}} Survey data such as \code{\link{contraceptive_use}}.}

\item{subnational}{'\emph{\sQuote{Logical}} If TRUE runs the sub national model.}

\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}
}
\value{
\emph{\sQuote{Data.frame}} Imputed survey data
}
\description{
If a user provides data where standard error (SE) columns have missing values, the 
corresponding pre-calculated SE from UNPD data are used to impute the missing values. 
The imputation of UNPD survey data is described at the bottom of this document in
the details section.
Imputation of (SE) variables is carried out annually with UNPD survey data \code{\link{fit_fp_csub}} 
as [@Cahill et al 2017 (appendix page 16)]. The summary of this procedure is described 
below. 
There are two scenarios. Each scenario has a corresponding procedure.
\describe{
    \item{1. (Completely missing) all entries missing for a coutnry's SE variable}
    \item{2. (Partially missing) some entries missing for a country's SE variable }
}
}
\details{
For scenario 1, the imputation is carried out by calculating the maximum of known 
sampling errors across all other countries and setting the unknown sampling errors 
equal to the median of these maximums.
For scenario 2 we impute the sampling errors by setting them equal to the maximum 
of the known sampling errors in that country.
}
