% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_from_global.R
\name{index_area}
\alias{index_area}
\title{index_area}
\usage{
index_area(meta, run_type)
}
\arguments{
\item{meta}{\emph{'Data.frame'} A data.frame with meta data from a global UNPD run}

\item{run_type}{\emph{'Character string'} "mwra" or "uwra"}
}
\value{
\emph{'Data.frame'} A data.frame with division data corresponding to a global run
}
\description{
Creates a data.frame similar to divisions data with codes used in the UNPD global run
}
