% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_from_global.R
\name{index_area_mwra}
\alias{index_area_mwra}
\title{index_area_mwra}
\usage{
index_area_mwra(meta)
}
\arguments{
\item{meta}{\emph{'Data.frame'} A data.frame with meta data from a global UNPD run}
}
\value{
\emph{'Data.frame'} A data.frame with division data corresponding to a global run
}
\description{
Creates a data.frame similar to divisions data with codes used in the UNPD global run
}
