% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_from_global.R
\name{index_datatype}
\alias{index_datatype}
\title{index_datatype}
\usage{
index_datatype(meta)
}
\arguments{
\item{meta}{\emph{'Data.frame'} A data.frame with meta data from a global UNPD run}
}
\value{
\emph{'List'}
\enumerate{
  \item \strong{map_cp}  \emph{'Data.frame'} A data.frame with source data corresponding to a global run for contracpetive data
  \item \strong{map_unmet}   \emph{'Data.frame'} A data.frame with source data corresponding to a global run for unmet data
}
}
\description{
Creates a list of data.frames for datatype codes used in the UNPD global run. These codes differ for contraceptive use and unmet data.
}
