% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_global_data.R
\name{list_global_data}
\alias{list_global_data}
\title{list_global_data
fit_fp_c
jags data obtained from the UNPD global model}
\usage{
list_global_data(is_in_union, core_data)
}
\arguments{
\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}

\item{core_data}{\emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link{core_data}}.}
}
\description{
list_global_data
fit_fp_c
jags data obtained from the UNPD global model
}
