% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples_all_women}
\alias{posterior_samples_all_women}
\title{Get posterior samples for all women}
\usage{
posterior_samples_all_women(
  in_union_posterior_samples,
  not_in_union_posterior_samples,
  core_data,
  population_data
)
}
\arguments{
\item{in_union_posterior_samples}{\emph{`array`} An array of n chains x n iterations x n years x n proportions}

\item{not_in_union_posterior_samples}{\emph{`array`} An array of n chains x n iterations x n years x n proportions}

\item{core_data}{consists of manipulated observations and run settings for \code{\link{fit_fp_c}}}

\item{population_data}{\emph{\sQuote{Data.frame}} Population count data such as \code{\link{population_counts}}.}
}
\value{
\emph{`array`} Posterior samples for all women
}
\description{
Combine posterior samples from in-union and not-in-union to obtain samples for all women
}
