% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples_array_format}
\alias{posterior_samples_array_format}
\title{Posterior samples reformat}
\usage{
posterior_samples_array_format(fit, core_data)
}
\arguments{
\item{fit}{the object returned from jags}

\item{core_data}{consists of manipulated observations and run settings for \code{\link{fit_fp_c}}}
}
\value{
\emph{\sQuote{Numeric array}} An array of samples of dimension chains x samples x years x proportions
}
\description{
Manipulate posterior samples into dimension chains x samples x years x proportions
}
