% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_transform.R
\name{proportions_vector}
\alias{proportions_vector}
\alias{modern_cpr}
\alias{traditional_cpr}
\alias{unmet_p}
\alias{total_cpr}
\alias{demand}
\alias{demand_modern}
\alias{no_need}
\alias{unmet_modern}
\alias{non_use}
\alias{demand_satisfied}
\alias{demand_satisfied_modern}
\title{Define three core proportions by order in the prortions vector.}
\usage{
modern_cpr(x)

traditional_cpr(x)

unmet_p(x)

total_cpr(x)

demand(x)

demand_modern(x)

no_need(x)

unmet_modern(x)

non_use(x)

demand_satisfied(x)

demand_satisfied_modern(x)
}
\arguments{
\item{x}{vector of at least length 3 with the first
three elements being: 1) modern CPR use
2) traditional CPR use; and 3) unmet need
for contraceptives}
}
\value{
proportion of in-universe respondents using
        a modern contraceptive method.

proportion of in-universe respondents using
        a traditional contraceptive method.

proportion of in-universe respondents with
        an unmet need for any contraceptive method

proprotion of in-universe respondents using
        either a modern or traditional contraceptive
        method.

proportion of in-uinverse respondents using
        a modern or traditional contraceptive method
        or in need of either method.

proportion of in-uinverse respondents using
        a modern contraceptive method
        or in need of a modern method.

proportion of in-universe respondents not
        using a contraceptive method and not in need
        of a contraceptive method.

proportion of in-universe respondents with
        an unmet need for a modern contraceptive method

proportion of in-universe respondents not
        using a modern or traditional contraceptive
        method.

proportion of demand for a
        contraceptive method satisifed using either a
        modern or traditional contracepritve method.

proprotion of demand among in-universe respondents for a
        contraceptive method satisifed using a
        modern contracepritve method.
}
\description{
Samples from the contraceptive use models always produce estimates
in order for a few basic proportions and we create accessors for these.
}
\details{
Further transformations refer to the accessors rather than using indexes
into the vector of proportions.
}
