% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_transform.R
\name{rotate_dimensions}
\alias{rotate_dimensions}
\title{rotate_dimensions}
\usage{
rotate_dimensions(x)
}
\arguments{
\item{x}{an array}
}
\value{
an array meeting the invariants
}
\description{
rotate the dimensions of an array s.t. the first margin becomes
the last.
}
\details{
The invariants for all arrays 'x' with more than one margin are:
dim(rotate_dimensions(x)) == c(dim(x)[2:D], dim(x)[1])
x['a','b','c'] == rotate_dimensions(x)['b','c','a']
}
