% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_summarise.R
\name{standard_quantiles}
\alias{standard_quantiles}
\title{Usual quantiles.}
\usage{
standard_quantiles()
}
\value{
Usual quantiles used to summarise credible intervals.
}
\description{
Usual quantiles.
}
