% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_summarise.R
\name{standard_summary}
\alias{standard_summary}
\title{Usual summary of a set of samples.}
\usage{
standard_summary(x, q = standard_quantiles())
}
\arguments{
\item{x}{set of parameter samples to summarize.}
}
\description{
Usual summary of a set of samples.
}
