% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_aggregate.R
\name{subset_population}
\alias{subset_population}
\title{subset_population}
\usage{
subset_population(population_data, divs_perlevel)
}
\arguments{
\item{population_data}{}

\item{divs_perlevel}{}
}
\value{

}
\description{
subset_population
}
