% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_aggregate.R
\name{subset_samples}
\alias{subset_samples}
\title{subset_samples}
\usage{
subset_samples(posterior_samples, divs_perlevel)
}
\arguments{
\item{posterior_samples}{}

\item{divs_perlevel}{}
}
\value{

}
\description{
subset_samples
}
