% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{to_json}
\alias{to_json}
\title{Convert data frame to JSON string}
\usage{
to_json(df)
}
\arguments{
\item{df}{`data.frame` Data frame to be converted}
}
\value{
`character` JSON string
}
\description{
Convert data frame to JSON string
}
