% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_transform.R
\name{transform}
\alias{transform}
\title{transform}
\usage{
transform(x, g, ...)
}
\arguments{
\item{x}{an array of samples with margins:
1) chain
2) iteration
4) year
5) three proportions in standard order:
     1) modern CPR
     2) traditional CPR
     3) unmet need for contraceptives}

\item{g}{a function that takes an array with years on the first margin
and a vector with at least three proportions in order on
the other: 1) modern CPR; 2) traditional CPR;
and 3) unmet need and returns one or more indicators.}
}
\value{
an array with the first two margins and the
        last margin modified to be 'indicator', which
        can be any transformation of the evolution of the three
        proprotions over the years..
}
\description{
Transform standard proprotions output to indicators
for single-unit array, summarizing years
}
