% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_transform.R
\name{transform_yearly}
\alias{transform_yearly}
\title{transform_yearly}
\usage{
transform_yearly(x, f, ...)
}
\arguments{
\item{x}{an array of samples with margins:
1) chain
2) iteration
4) year
5) three proportions in standard order:
     1) modern CPR
     2) traditional CPR
     3) unmet need for contraceptives}

\item{f}{a function that takes a vector with at least three
proportions in order: 1) modern CPR; 2) traditional CPR;
and 3) unmet need and returns one or more indicators.}
}
\value{
an array with the first three margins and the
        last margin modified to be 'indicator', which
        can be any transformation of the three proprotions.
}
\description{
Transform standard proprotions output to indicators
for single-unit array, applied yearly
}
