% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_aggregate.R
\name{weight_input_checker}
\alias{weight_input_checker}
\title{weight_input_checker}
\usage{
weight_input_checker(division_level_data, population_data, posterior_samples)
}
\value{
informative error messages when inputs are incorrect
}
\description{
weight_input_checker
}
