% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_sequence_maker.R
\name{year_sequence_maker}
\alias{year_sequence_maker}
\title{year_sequence_maker}
\usage{
year_sequence_maker(first_year_max, first_year, last_year, contraceptive_use)
}
\arguments{
\item{first_year_max}{model estimation cannot start any later than this year}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}

\item{last_year}{\emph{\sQuote{Numeric}} The last year of model estimates in output. The model will be fit to all data, including dates after this date if available.}

\item{contraceptive_use}{}
}
\value{
\emph{\sQuote{List}}
\enumerate{
  \item \strong{model_seq}  \emph{\sQuote{Numeric vector}} Year index for model estimation
  \item \strong{model_seq_years}  \emph{\sQuote{Numeric vector}} Years of model estimation
  \item \strong{result_seq }  \emph{\sQuote{Numeric vector}} Year index for results based on user request
  \item \strong{result_seq_years}  \emph{\sQuote{Numeric vector}} Years of results based on user request
}
}
\description{
creates a list of years and respective indices based on user specified years to estimate and observed years
}
