% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bias_data.R
\name{list_bias_data}
\alias{list_bias_data}
\title{list_bias_data}
\usage{
list_bias_data(core_data)
}
\arguments{
\item{core_data}{\emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link{core_data}}.}
}
\description{
jags data, numeric form of bias using \code{\link{core_data}}.
}
